Add rubocop config, rubocop autocorrect

+++ db/migrate/20171219086058_add_index_to_spree_stock_items_variant_id.spree.rb
@@ -1,14 +1,10 @@
 # This migration comes from spree (originally 20150216173445)
 class AddIndexToSpreeStockItemsVariantId < ActiveRecord::Migration[4.2]
   def up
-    unless index_exists? :spree_stock_items, :variant_id
-      add_index :spree_stock_items, :variant_id
-    end
+    add_index :spree_stock_items, :variant_id unless index_exists? :spree_stock_items, :variant_id
   end
 
   def down
-    if index_exists? :spree_stock_items, :variant_id
-      remove_index :spree_stock_items, :variant_id
-    end
+    remove_index :spree_stock_items, :variant_id if index_exists? :spree_stock_items, :variant_id
   end
 end
