+++ db/migrate/20171219086062_remove_user_index_from_spree_state_changes.spree.rb
@@ -1,15 +1,10 @@
 # This migration comes from spree (originally 20150324104002)
 class RemoveUserIndexFromSpreeStateChanges < ActiveRecord::Migration[4.2]
   def up
-    if index_exists? :spree_state_changes, :user_id
-      remove_index :spree_state_changes, :user_id
-    end
-
+    remove_index :spree_state_changes, :user_id if index_exists? :spree_state_changes, :user_id
   end
 
   def down
-    unless index_exists? :spree_state_changes, :user_id
-      add_index :spree_state_changes, :user_id
-    end
+    add_index :spree_state_changes, :user_id unless index_exists? :spree_state_changes, :user_id
   end
 end
