+++ db/migrate/20171219086077_remove_counter_cache_from_spree_variants_to_spree_stock_items.spree.rb
@@ -1,11 +1,8 @@
 # This migration comes from spree (originally 20160509064646)
 class RemoveCounterCacheFromSpreeVariantsToSpreeStockItems < ActiveRecord::Migration[4.2]
   def up
-    if column_exists?(:spree_variants, :stock_items_count)
-      remove_column :spree_variants, :stock_items_count
-    end
+    remove_column :spree_variants, :stock_items_count if column_exists?(:spree_variants, :stock_items_count)
   end
 
-  def down
-  end
+  def down; end
 end
